moved ActionDelegate to delegates.py
